Moved the Entity classes to respective sub-packages
