+++ src/main/java/in/gagan/base/framework/entity/user/User.java
@@ -17,7 +17,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package in.gagan.base.framework.entity;
+package in.gagan.base.framework.entity.user;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
