+++ src/main/java/in/gagan/base/framework/service/user/UserDataServiceImpl.java
@@ -31,7 +31,7 @@ import org.springframework.stereotype.Service;
 
 import in.gagan.base.framework.component.PasswordProps;
 import in.gagan.base.framework.dao.user.UserDAO;
-import in.gagan.base.framework.entity.User;
+import in.gagan.base.framework.entity.user.User;
 import in.gagan.base.framework.util.ExceptionHelperUtil;
 
 /**
