+++ src/main/java/in/gagan/base/framework/service/user/UserRegisterationServiceImpl.java
@@ -29,8 +29,8 @@ import org.springframework.stereotype.Service;
 import in.gagan.base.framework.constant.ApplicationConstants;
 import in.gagan.base.framework.dto.user.UpdateUserDTO;
 import in.gagan.base.framework.dto.user.UserDTO;
-import in.gagan.base.framework.entity.User;
-import in.gagan.base.framework.entity.VerificationToken;
+import in.gagan.base.framework.entity.user.User;
+import in.gagan.base.framework.entity.user.VerificationToken;
 import in.gagan.base.framework.exception.UsernameExistException;
 import in.gagan.base.framework.util.UserHelperUtil;
