+++ src/main/java/in/gagan/base/framework/util/DTOMapper.java
@@ -2,8 +2,8 @@ package in.gagan.base.framework.util;
 
 import in.gagan.base.framework.dto.user.UserDTO;
 import in.gagan.base.framework.dto.user.UserRoleDTO;
-import in.gagan.base.framework.entity.Role;
-import in.gagan.base.framework.entity.User;
+import in.gagan.base.framework.entity.user.Role;
+import in.gagan.base.framework.entity.user.User;
 import org.springframework.beans.BeanUtils;
 
 import java.util.Set;
